# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Arcana API Gateway
  version: 0.1.0
  description: |
    Unified API for the Arcana AI platform. The gateway exposes native routes for
    platform health, chat, authentication, and enterprise features, and reverse-proxies
    requests to backend services across eight planes (agent, data, tool, model, govern,
    quality, ops, infra).
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.arcana.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/eval/run:
    post:
      tags: [Quality Plane]
      summary: Run skill evaluation (Probe)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                skill_ref:
                  type: string
                judges:
                  type: array
                  items:
                    type: string
      responses:
        '202':
          description: Evaluation started
...
  /api/v1/catalog/skills:
    post:
      tags: [Ops Plane]
      summary: Register skill in catalog (Registry)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                version:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Skill registered
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued via /api/v1/auth/keys
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from /api/v1/auth/token or external IdP
...